Make BootstrapHandler as an interface and add DefaultBootstrapHandler
